+++ src/main/java/ru/saiev/interviewtask/controller/StudentController.java
@@ -21,6 +21,10 @@ public class StudentController {
     public ResponseEntity<StudentDto> createStudent(@RequestBody StudentDto student) {
         StudentDto savedStudent = studentService.createStudent(student);
 
+        if (studentService.getStudentById(student.getId()) != null) {
+            throw new StudentNotFoundException("Студент с такими данными уже существует!");
+        }
+
         return new ResponseEntity<>(savedStudent, HttpStatus.CREATED);
     }
 
@@ -38,12 +42,18 @@ public class StudentController {
     @GetMapping
     public ResponseEntity<List<StudentDto>> getAllStudents() {
         List<StudentDto> students = studentService.getAllStudents();
+        if (students.isEmpty()) {
+            throw new StudentNotFoundException("Список студентов пуст!");
+        }
 
         return new ResponseEntity<>(students, HttpStatus.OK);
     }
 
     @PutMapping("{id}")
     public ResponseEntity<StudentDto> updateStudent(@PathVariable("id") Long studentId, @RequestBody StudentDto studentDto) {
+        if (studentService.getStudentById(studentId) == null) {
+            throw new StudentNotFoundException("Не найден студент с таким Id!" + studentId);
+        }
         studentDto.setId(studentId);
         StudentDto updatedStudent = studentService.updateStudent(studentDto);
 
@@ -52,6 +62,10 @@ public class StudentController {
 
     @DeleteMapping("{id}")
     public ResponseEntity<String> deleteStudent(@PathVariable("id") Long studentId) {
+        if (studentService.getStudentById(studentId) == null) {
+            throw new StudentNotFoundException("Не найден студент с таким Id!" + studentId);
+        }
+
         studentService.deleteStudent(studentId);
 
         return new ResponseEntity<>("Студент успешно удалён!", HttpStatus.OK);
